Adding PIX events and object names to samples
Demonstrating how to use event markers and the DX naming API to improve
the graphics debugging experience.  Use ALT+F5 in Visual Studio to
launch the graphics debugger then press the PrtScn key to capture
rendered frames that can be played back and inspected for debugging
purposes.

Hello* is not instrumented to keep that particular sample series simple.
Multiadapter and ReservedResources are not instrumented due to the
debugger not supporting those scenarios at this time.