Account for adapters that don't support cross-adapter row-major texture sampling
This sample was originally written with hybrid laptops in mind.  The
Intel integrated GPUs in these laptops support sampling directly from
the shared texture, so this feature cap was not originally queried.  The
sample has been updated to query for this capability and provide an
example of how to share resources across adapters in other types of
multiadapter systems.