Fixes to incorrect Shader Record offset calculations, partial support for Root Views, and Root Signature Serialization fixes (#370)
Big chunk of Fallback Layer fixes/features:
Rev'ing the Fallback Layer to work with a newer version of DxCompiler. This fixes multiple issues including invalid shader record alignment calculations and handling for multiple descriptor tables in a shader record that was causing failures to link
New DxCompiler version enables partial support for root views, but offset in bytes is still not supported. To prevent use without knowledge of the short-comings, shielding the use of root descriptors behind an explicit flag to stop developers from using this without opt-in
Fix to root signature serialization and unittests to catch future regressions
ModelViewer had an invalid shader record offset calculation that was revealed after the DxCompiler fixes