User VS HLSL compiler build step, use dxc binaries from SDK. (#466)
Removed dependency on private binaries of dxc, dxcompiler and dxil. The ones from SDK are used.
DXC is now supported as of VS 15.8. With that:
--- Switched from custom build step to VS HLSL compiler build step for shaders.
--- Removed dependency on a hack of renaming dxc to fxc. This is not needed anymore as VS now uses dxc for SM 6.0+. Moved Compute Shaders to SM 6.0 so that dxc is used.
Cleaned up FL and FLunittests vcxproj. There were a lot of incorrect or inconsistent settings, i.e. use of SM 4.0, which were ignored previously, because dxc was being forced.