Conditional use of d3d12.1 features
d3dx12 contains features that require the inclusion of the d3d12_1.h header, which is not present on every (any ?) version of the Windows SDK - at least not under the v1909 one. 

This change includes a preprocessor condition in order to include d3d12_1.h if possible when compiling in a C++17 environment, otherwise it will only enable DX12.1 features if d3d12_1.h has been included before d3dx12.