Fix recognition of WM_QUIT (#776)
If multiple messages are received at the same time as the quit message, then we can miss it because we only check the last message received. To repro this, alt-F4 while moving the mouse. Fix by just keeping track of whether we ever saw WM_QUIT in the message dispatch loop.